Update js file to modules
