+++ about.html
@@ -146,6 +146,6 @@
       </div>
     </footer>
 
-    <script src="about.js"></script>
+    <script type="module" src="about.js"></script>
   </body>
 </html>
