+++ obj.js
@@ -1,0 +1,40 @@
+const speakers = [
+  {
+    name: 'Vitalik Buterin',
+    carrier: 'Berkman Professor of Enterpreneurial legal Studies',
+    bio: 'Benkler studies commons-based peer production, and published his seminar book. The wealth of Networks',
+    image: 'asset/vitalik.jpg',
+  },
+  {
+    name: 'Anthony Pompliano',
+    carrier: 'Director of Art Centre Nabi and a board member CC Korea',
+    bio: 'As the main venue for new media art production in Korea,Nabi promotes cross-disciplinary collaboration.',
+    image: 'asset/anthony.jpg',
+  },
+  {
+    name: 'Charlie Lee ',
+    carrier: 'Executive Diretor of the Wikimedia Foundation',
+    bio: 'Lila Tretikov is the executive Director of Wikipedia Foundation, the nonprofit organization that operates Wikipedia.',
+    image: 'asset/lee.jpg',
+  },
+  {
+    name: 'Tim Draper',
+    carrier: 'Executive Diretor of the Wikimedia Foundation',
+    bio: 'Kilnam Chon helped bring the internet to asia and is an outspoken advocate for the open web and digital commons.',
+    image: 'asset/tim.jpg',
+  },
+  {
+    name: 'Roger Ver',
+    carrier: 'President of Young Pirates of Europe',
+    bio: 'European ingetration, political democracy and participation of youth through online as her major condern.',
+    image: 'asset/roger.jpg',
+  },
+  {
+    name: 'Andreas M. Antonopoulos',
+    carrier: 'CEO of Creative, Commons, ex CEO of the Mozilla Foundation',
+    bio: 'Ryan had been leading open-source projects at the Mozilla Foundation such as the open source movement',
+    image: 'asset/andreas.jpg',
+  },
+];
+
+export default speakers;
